EM-2945: Move swagger file into main folder
